styling pending boxes

+++ backend-flask/app/models/transaction.py
@@ -10,7 +10,7 @@ class Transaction(db.Model):
     payee_id = db.Column(db.Integer, db.ForeignKey("users.id"), nullable=False)
     creator_id = db.Column(
         db.Integer, db.ForeignKey("users.id"), nullable=False)
-    amount = db.Column(db.Float(2), nullable=False)
+    amount = db.Column(db.Float(), nullable=False)
     details = db.Column(db.Text, nullable=False)
     paid = db.Column(db.Boolean, nullable=False)
     created_at = db.Column(db.DateTime, server_default=func.now())
